pyproject.toml
# pyproject.toml
#
# This is the main configuration file for the aicli project. It uses Poetry for dependency
# management and packaging. Poetry is a modern Python packaging tool that simplifies
# dependency management and package distribution.

[tool.poetry]
# Basic project metadata
name = "aicli"                                  # Package name as it will appear in PyPI
version = "0.1.0"                               # Semantic versioning
description = "AI-powered command line assistant"
authors = ["Your Name <your.email@example.com>"]

# Project dependencies
[tool.poetry.dependencies]
python = "^3.8"                                 # Requires Python 3.8 or higher
click = "^8.1.7"                               # CLI framework for building command line tools
openai = "^1.12.0"                             # OpenAI API client for AI interactions
pyyaml = "^6.0.1"                              # YAML parser for configuration files
python-dotenv = "^1.0.0"                       # Environment variable management

# CLI entry point configuration
[tool.poetry.scripts]
aicli = "aicli.main:cli"                       # Maps 'aicli' command to main.py's cli function

# Build system configuration
[build-system]
requires = ["poetry-core>=1.0.0"]              # Poetry build requirements
build-backend = "poetry.core.masonry.api"      # Specifies Poetry as the build system